test(orchestration): split runner.test.ts into six topic-coherent files
runner.test.ts had grown to 3868 lines with 35 describe blocks and 184
tests across six unrelated topic areas. Split into:

- runner.lifecycle.test.ts — turn/agent lifecycle primitives (orchOnStop,
  refreshAgentStatuses, marker files, crash recovery, snapshot reconcile)
- runner.hung-agents.test.ts — stall detection and nudging
- runner.pr-comments.test.ts — PR-comment monitoring and review plumbing
- runner.auto-commit.test.ts — auto-commit phase side effects
- runner.verification.test.ts — verify/validate phase gates
- runner.phase-skipping.test.ts — phase-transition plumbing

Extract nine shared helpers (makeTmpDir, makeGitRepo, makeLifecycle,
makeState, markAgentDone, makePeerSyncDir, makeSnapshot, makeMockTransport,
noopTransport) plus the gh-ludics-199 "Testing pattern" banner into a new
runner.test-helpers.ts (not .test.ts — bun's glob would pick it up).
Auto-commit-only helpers stay inline in runner.auto-commit.test.ts.

docs/testing-patterns.md pointers updated to the new layout. No change to
src/orchestration/runner.ts itself.

Test count preserved: 184 tests, 371 expect() calls, all passing.
Full suite: 1070/1070. Typecheck, build, and CI lint scripts all green.
